mybarter/requirements.txt
Django==5.2a1
djangorestframework==3.15.2
django-filter==25.1
drf-yasg==1.21.10
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.1